fix rendering of model IDs with a colon in their name (#704)
